requirements.txt
pypdf
langchain
torch
accelerate
bitsandbytes
transformers
sentence_transformers
faiss_cpu
streamlit-chat
